# api/openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce API
  version: "1.0.0"
  description: |
    HTTP API consumed by the React storefront. All responses follow the `{ "data": ... }` envelope on success and `{ "error": { message, status } }` on failure.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            status:
              type: integer
            path:
              type: string
            meta:
              type: object
          required:
            - message
            - status
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - email
        - name
...
    AuthCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
...
    AuthEnvelope:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
          required:
            - user
      required:
        - data
    OkEnvelope:
      type: object
      properties:
        data:
          type: object
          properties:
            ok:
              type: boolean
          required:
            - ok
      required:
        - data
...
paths:
...
  /auth/signin:
    post:
      summary: Authenticate an existing user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        "200":
          description: Signed in
          headers:
            X-Access-Token:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: Refresh access token using the refresh cookie
      tags: [Auth]
      responses:
        "200":
          description: Access token rotated
          headers:
            X-Access-Token:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        "401":
          description: Missing or invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...